MAINT: stats.fit: move to new file
