+++ scipy/stats/_result_classes.py
@@ -24,4 +24,4 @@ __all__ = ['BinomTestResult', 'RelativeRiskResult', 'TukeyHSDResult',
 from ._binomtest import BinomTestResult
 from ._relative_risk import RelativeRiskResult
 from ._hypotests import TukeyHSDResult
-from ._distn_infrastructure import FitResult
+from ._fit import FitResult
